requirements.txt
GitPython
tree_sitter
networkx
matplotlib
python-ternary
stop-words
nltk
gensim
python-Levenshtein
ipywidgets
tqdm
pyfiglet
numpy
biterm
jupyterlab
regex  # this one, as opposed to the standard re module, can also split strings with zero width patterns, which this project requires
scipy
# jupyter nbextension enable --py widgetsnbextension

ipython
cloudpickle
dill
pprofile
mtprof
